Invite all team members to common channels
This is currently implemented by setting the default channels for the
usergroups. This works because every team member is in a usergroup and
usergroups have the functionality of ensuring channel presence.

In the future this might have to be implemented more manually, like the channel
presence for the main team channels is currently implemented.