Change <team-name> group name to <team-name>-team
This is essentially to allow the team name to use _reserved_ names, such as
`here` or `channel`, which are reserved to trigger specific notifications.

With channel names, we use a prefix, because a shared prefix is important for
grouping all of the created channels together in Slack's left-hand sidebar.
Groups are very rarely browsed and so such alphabetical ordering doesn't play
a role and so the group names do not need to share a prefix. In theory
`team-<team-name>` would also be available, but I think that would just cause
confusion between the `tm` and `team` prefixes.