Support private team channels
Currently, `findChannel` would not find private channels even if the bot has
the necessary oAuth scope permissions and is invited to that channel. The code
would then try to create that channel instead but would fail due to the name
conflict.

This change makes private channels visible to the `findChannel` method,
allowing private channels to be used by the bot, if necessary permissions are
in place and the bot is invited to that private channel.