Fall back to user's name if display name is not set
Currently, for many users, the bot sets their `_displayName` as just "@".
Although group assignments etc work fine, then this can be very confusing for
the user in channel topics etc.

It happens because having a display name is not mandatory for Slack users. So
for users who haven't configured it, it's not present. The code checks for both
presence and emptiness, falling back to just the "name" property if the display
name is not present or is empty.