Query more Slack channels at a time
Unfortunately there isn't an API for looking up a channel by name, so what we
do is list all of the channels and see if there are any with the name we're
looking for. Unfortunately, doing this again, and again for every team,
including multiple times for each page of the results, we can sometimes get
rate-limited.

This changes the amount of channels returned in a single page from the default
of 100 to the maximum 1000. This should decrease the amount of requests
required and should minimize the risk of getting rate-limited. See [the
endpoint docs][1] for the default and max information.

Also considered caching the results, because we don't really need to pull all
of the channels for each team. However, I wasn't able to get Wreq's built-in
cache support working, so I abandoned the idea for now. Didn't want to manually
implement it either, if a general HTTP cache implementation could cover it.
Ideally the unnecessary queries could still be cut in the future.

[1]: https://api.slack.com/methods/conversations.list