Remove support for the global caretakers group
With many different teams, this group becomes too big to be usable. Some of the
teams are likely to be more closely related than other, so sending a message to
the whole group starts becoming irrelevant for some participants.

There is still value in cross-team groups, but there will have to be a more
flexible way of managing them. Just including all caretakers from all teams
doesn't work.