Move some pure logic to a separate Config module
This is still mostly duplicated in the `Lib` module, but the duplication can
now be removed, step by step. The main benefit of this is that the logic now
becomes more easily testable, as shown by the one added test. Additional tests
can and will be added shortly as well, once the `Lib` module has been
refactored to make more use of this logic.