Request user information only once per execution
Previously many requests were made to Slack APIs to get user information. Now
the user information is requested once, stored in a map, and then requested
locally from that map if required.

For the actual `ensure` run, this was executed only for evaluating the topic
message for each channel. For the `--dry-run` however, it was executed many
more times, making the process very slow.