Introduce polysemy
These first steps are roughly inspired by the [Porting to Polysemy][1] article.

The `GetDisplayName` approach is the common work-around for impredicative
types, explained e.g. in [this SO post][2]. There is a new implementation of
the `ImpredicativeTypes` language extension starting from GHC 9.2.1 which would
avoid the need for doing this. However, the current stackage snapshot this
project is on uses ghc-8.10.7 and version 9 isn't supported in any of the LTS
snapshots yet. Will have to wait a bit before I'd be able to use it here.

[1]: https://reasonablypolymorphic.com/blog/porting-to-polysemy/
[2]: https://stackoverflow.com/a/56449258/1456578