Move NetCtx to an effect
Now it's still threaded through the application in types but no longer needs to
be passed through everywhere in the implementation.

To be honest, I'm not certain about how I've defined `runNetCtx`. It seems to
work and only create the session once but I'm not sure if that's accidental or
if it actually works.