Only query the channel list once
Previously the [conversations.list endpoint][1] was queried for each
`Channel.find`, which means it was queried once for each team. The
endpoint has a [Tier 2][2] rate limit, which is described as "20+ per
minute" which sets a fairly strict limit on the number of teams that
could be managed by this.

[1]: https://api.slack.com/methods/conversations.list
[2]: https://api.slack.com/docs/rate-limits#tier_t2