Implement automatic caching for users
Previously the caller had to explicitly make a single `Users.listAll`
call and then re-use the response to avoid spamming the Slack API with
unnecenssary requests. Now, the caching happens automatically in the
background, so the caller can use `Users.listAll` and `Users.find`
directly as much as needed, without worrying about spamming the APIs.