Remove text output from ensure method
This fixes the last test that has been failing since the beginning of
the large refactoring. The test has been failing as a reminder to
restore "Team x: Success!" functionality in the final output. However,
there's no concept of a _team_ for the configuration that gets fed to
the program now. Because of this, the functionality has instead been
replaced with proper logging support.