Freeze and cache zoo dependencies
This expects all the files in `./types/zoo` to already be "frozen"
(imports have hashes) before the Docker build. The build fails if they
are not frozen. Additionally, the `dhall freeze --check` also populates
the local cache, which can then be made available in the runtime image
as well, avoiding unnecessary downloads.