Fix dhall freeze --check in Docker build
The previous version had two issues. First, the the globstar shell
option is not enabled by default for `RUN` and so the `**` didn't work
as expected. Second, even if it did work, it would also include
sub-folders in the list, which `dhall freeze --check` does not accept.

A simple `find` describes more explicitly which files need to be
checked and doesn't have these problems.