Add deduplication logic to avoid reverting manual changes
For inflexible rotations, such as `Weekly`, users frequently employ
manual overrides for flexibility. For example, when a caretaker is out
sick on a particular day, then Slack user groups and channel topics can
be updated manually for someone else to take over the caretaking
responsibilities for the time being.

The goal of these changes is to respect those manual changes when
re-applying configuration that has already been applied before. For
example, when running the bot every hour to get the latest schedule from
Opsgenie, we want to avoid reverting manual changes made to less
flexible rotations (such as `Weekly`).

The solution here is to store information on disk about changes that
have already been applied and check against that store when deciding
whether to apply the same configuration again during a subsequent
execution. SQLite is used for storing and querying the data.