Change deduplicator to apply for all rotations
The main benefit of this is that this allows the script to be run more
frequently. Even running every 5 to 10 minutes should no longer be a
problem, because unless the "output" of any rotation changes, we don't
need to call Slack to check for current status of things, etc.

The only API queries that are made on every execution are the ones made
to e.g. Opsgenie to resolve the rotation. Opsgenie (per account) rate
limits start from 100 per second and 500 per minute, so doing as many
requests as there are rotations every 5 minutes should not be a problem.

This change also gets rid of the `nextWeek` logic that's specific to the
`Weekly` rotation. It just waits for the output to change, whatever the
cause, and decides based on that. This works for all types of rotations.

Detecting "stale" data is now a bit harder but the implemented solution
that just deletes data that hasnt been touched in 14 days is a simple
solution that shouldn't cause any problems and should keep the size
under control as well.

[1]: https://docs.opsgenie.com/docs/api-rate-limiting#rate-limit-calculation