Retry Slack request when hitting rate limits
As described in the changelog, this trades unnecessary failures in favor
of some delay. In most cases a delay of up to a few minutes is less of
a problem than full or partial failure.

The retry limit of 4 is arbitrary. In most cases one should be enough,
but with multiple concurrent invocations (e.g. by different people using
the same token for --dry-run) more may be necessary. 4 was chosen as
something that should let each request likely to pass eventually without
leaving the bot in a bad state for too long.

The ratelimiting behavior is confirmed in practice and also documented
in [Slack API docs][1].

[1]: https://api.slack.com/docs/rate-limits#headers